[net] sched: cls_flower: fix infinite loop in fl_walk()
Message-id: <63029fdfcea7c0fe98d8402779dc498b8f109256.1561375326.git.dcaratti@redhat.com>
Patchwork-id: 265807
O-Subject: [RHEL8.1 net PATCH] net/sched: cls_flower: fix infinite loop in fl_walk()
Bugzilla: 1712739
RH-Acked-by: Paolo Abeni <pabeni@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1712739
Upstream Status: RHEL-Only
Brew:https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=22321515
Tested: add a flower rule with handle 0xffffffff and then read it back

after dumping a flower rule having 'handle' equal to 0xffffffff, fl_walk()
increases the value of 'cookie' and iterates over the next idr. Assigning
f->handle + 1 causes an infinite loop, because 'handle' is u32 and '0' is
a cookie value for which tcf_fill_node() still dumps the current rule.

Since 'cookie' is unsigned long, we can fix 64-bit architectures just with
a cast to unsigned long for 'handle', before doing the increase, so that
the integer overflow does not occur.
Please note:

1) on 64-bit CPUs running upstream Linux, the problem is not visible
   anymore since commit 061775583e35 ("net: sched: flower: introduce
   reference counting for filters") does arg->cookie++, and does not
   overflow. This commit was part of the RTNL lock removal, it had no
   "Fixes:" tag and so it wasn't backported at all.
2) on 32 bit CPUs running Linux the problem is still present (I verified
   it on 5.2.0-0.rc5.git0.1.fc31.i686).

A follow-up patch targeting Dave Miller's "net" tree that provides a fix
for 32-bit CPUs, and a TDC selftest that verifies the correct behaviour,
will then be sent.

+++ net/sched/cls_flower.c
@@ -1470,7 +1470,7 @@ static void fl_walk(struct tcf_proto *tp, struct tcf_walker *arg,
 			arg->stop = 1;
 			break;
 		}
-		arg->cookie = f->handle + 1;
+		arg->cookie = (unsigned long)f->handle + 1;
 		arg->count++;
 	}
 }
